@@ -1,6 +1,6 @@
 {
   "name": "ng2-intl-tb",
-  "version": "1.0.5",
+  "version": "1.0.6",
   "private": true,
   "description": "Internationalize Angular2 apps. This library provides Angular2 components, pipe and an API to format dates, numbers, and strings, including pluralization and handling translations.",
   "keywords": [
@@ -44,28 +44,28 @@
   },
   "homepage": "https://github.com/eyolas/ng2-intl#readme",
   "dependencies": {
-    "intl-format-cache": "^2.0.5",
-    "intl-messageformat": "^2.0.0",
-    "intl-relativeformat": "^2.0.0",
-    "lodash": "^4.16.6"
+    "intl-format-cache": "^2.1.0",
+    "intl-messageformat": "^2.2.0",
+    "intl-relativeformat": "^2.1.0",
+    "lodash": "^4.17.10"
   },
   "peerDependencies": {
-    "@angular/core": "^4.0.0",
-    "@angular/common": "^4.0.0",
-    "rxjs": "^5.0.1"
+    "@angular/core": "^6.0.0",
+    "@angular/common": "^6.0.0",
+    "rxjs": "^6.0.0"
   },
   "devDependencies": {
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/compiler-cli": "^4.0.0",
-    "@angular/core": "^4.0.0",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@types/jasmine": "^2.5.38",
-    "@types/lodash": "^4.14.38",
-    "@types/node": "^6.0.79",
+    "@angular/common": "^6.0.0",
+    "@angular/compiler": "^6.0.0",
+    "@angular/compiler-cli": "^6.0.0",
+    "@angular/core": "^6.0.0",
+    "@angular/platform-browser": "^6.0.0",
+    "@angular/platform-browser-dynamic": "^6.0.0",
+    "@types/jasmine": "^2.8.7",
+    "@types/lodash": "^4.14.109",
+    "@types/node": "^6.0.111",
     "awesome-typescript-loader": "3.0.0-beta.10",
-    "codelyzer": "^3.0.0",
+    "codelyzer": "^3.2.2",
     "cz-conventional-changelog": "^1.2.0",
     "formatjs-extract-cldr-data": "^3.0.0",
     "istanbul-instrumenter-loader": "0.2.0",
@@ -74,26 +74,25 @@
     "karma-chrome-launcher": "2.0.0",
     "karma-coverage": "1.1.1",
     "karma-jasmine": "1.0.2",
-    "karma-mocha-reporter": "^2.1.0",
-    "karma-remap-coverage": "~0.1.2",
+    "karma-mocha-reporter": "^2.2.5",
+    "karma-remap-coverage": "^0.1.5",
     "karma-sourcemap-loader": "^0.3.7",
-    "karma-webpack": "^2.0.0",
+    "karma-webpack": "^2.0.13",
     "mkdirp": "^0.5.1",
-    "rimraf": "^2.5.4",
+    "rimraf": "^2.6.2",
     "rollup": "^0.36.3",
     "rollup-plugin-memory": "^2.0.0",
     "rollup-plugin-uglify": "^1.0.1",
-    "rxjs": "^5.0.1",
-    "serialize-javascript": "^1.3.0",
+    "serialize-javascript": "^1.5.0",
     "shx": "^0.2.2",
     "source-map-loader": "0.1.5",
     "ts-helpers": "1.1.2",
     "ts-node": "^3.1.0",
-    "tslint": "^5.0.0",
-    "tslint-loader": "^3.5.3",
-    "typescript": "^2.4.0",
+    "tslint": "^5.10.0",
+    "tslint-loader": "^3.6.0",
+    "typescript": "2.7.*",
     "webpack": "^2.4.1",
-    "zone.js": "^0.8.4"
+    "zone.js": "^0.8.26"
   },
   "optionalDependencies": {
     "debug": "^2.2.0"
