@@ -1,8 +1,7 @@
 import IntlRelativeFormat from 'intl-relativeformat';
 import * as formatters from '../formatters';
 import { escape, filterProps } from '../utils';
-import { Observable } from 'rxjs/Observable';
-import { Observer } from 'rxjs/Observer';
+import { Observable, Observer } from 'rxjs';
 import { Injectable } from '@angular/core';
 import { IntlService } from './intl.service';
 import {
@@ -191,7 +190,7 @@ export class FormatService {
         observer.complete();
     });
   }
-  
+
   formatText(descriptor: MessageDescriptor, values = {}): string {
     let {
       id
